Build program with static link.

+++ convert/convertnumber.h
@@ -11,7 +11,7 @@ public:
     ConvertNumber(){}
     ~ConvertNumber(){}
     string numberToString(int value){
-        return std::to_string(value);
+        return std::to_string(static_cast<long long>(value));
     }
     int stringToNumber(string value){
         return std::stoi(value);
